20/10/2021

+++ ft_bzero.c
@@ -3,17 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kevyn <marvin@42.fr>                       +#+  +:+       +#+        */
+/*   By: kcatrix <kcatrix@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/15 13:34:29 by kevyn             #+#    #+#             */
-/*   Updated: 2021/10/15 16:43:06 by kevyn            ###   ########.fr       */
+/*   Updated: 2021/10/20 17:00:57 by kcatrix          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <string.h>
+#include <stdio.h>
+
 void	ft_bzero(void *s, size_t n)
 {
-	int		i;
-	char	*str;
+	size_t		i;
+	char		*str;
 
 	str = s;
 	i = 0;
