+++ ft_strlcpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kcatrix <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: kcatrix <kcatrix@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/18 13:31:00 by kcatrix           #+#    #+#             */
-/*   Updated: 2021/10/19 11:20:59 by kcatrix          ###   ########.fr       */
+/*   Updated: 2021/10/20 17:04:41 by kcatrix          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *restrict dst, const char *restrict src, size_t d)
 {
-	int	i;
-	int	len;
+	size_t	i;
+	int		len;
 
 	i = 0;
 	len = 0;
